requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
rq==1.16.2
redis==5.0.8
requests==2.32.3
pdf2image==1.17.0
Pillow==10.4.0
python-dotenv==1.0.1
